% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-block-output.R
\name{block_output.final_block}
\alias{block_output.final_block}
\title{Custom Output Method for Final Block}
\usage{
\method{block_output}{final_block}(x, result, session)
}
\description{
Displays dygraph with optional outlier annotations
}
