% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.seasonal-package}
\alias{blockr.seasonal}
\alias{blockr.seasonal-package}
\title{blockr.seasonal: Seasonal Adjustment Blocks for blockr}
\description{
A blockr package for seasonal adjustment using the seasonal package. Provides blocks for seasonal adjustment of time series data using X-13ARIMA-SEATS and TRAMO/SEATS methods via the seasonal package.
}
\author{
\strong{Maintainer}: Christoph Sax \email{christoph@cynkra.com}

}
\keyword{internal}
