% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal-block.R
\name{generate_sample_data}
\alias{generate_sample_data}
\title{Generate Sample Data (Internal Helper)}
\usage{
generate_sample_data(dataset_name, n_rows, add_noise = FALSE)
}
\arguments{
\item{dataset_name}{Name of the dataset}

\item{n_rows}{Number of rows to sample}

\item{add_noise}{Whether to add noise to numeric columns}
}
\description{
Internal function to generate sample data based on parameters.
}
\keyword{internal}
