% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-block.R
\name{html_coefs}
\alias{html_coefs}
\title{Generate HTML coefficients display (seasonalview-style)}
\usage{
html_coefs(x)
}
\arguments{
\item{x}{A seas model object}
}
\value{
HTML content for coefficients
}
\description{
Generate HTML coefficients display (seasonalview-style)
}
