% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-block.R
\name{html_stats}
\alias{html_stats}
\title{Generate HTML statistics display (seasonalview-style)}
\usage{
html_stats(x, digits = 5)
}
\arguments{
\item{x}{A seas model object}

\item{digits}{Number of digits for formatting}
}
\value{
HTML content for statistics
}
\description{
Generate HTML statistics display (seasonalview-style)
}
