% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-block.R
\name{html_summary}
\alias{html_summary}
\title{Generate HTML summary for seas model (main function)}
\usage{
html_summary(seas_model)
}
\arguments{
\item{seas_model}{A seas model object}
}
\value{
A tagList with HTML content
}
\description{
Pure function to generate HTML display from a seas model
}
