% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-block.R
\name{html_tests}
\alias{html_tests}
\title{Generate HTML tests display (seasonalview-style)}
\usage{
html_tests(x, digits = 4)
}
\arguments{
\item{x}{A seas model object}

\item{digits}{Number of digits for formatting}
}
\value{
HTML content for tests
}
\description{
Generate HTML tests display (seasonalview-style)
}
