% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fivebestmdl-block.R
\name{new_fivebestmdl_block}
\alias{new_fivebestmdl_block}
\title{Five Best Models Block}
\usage{
new_fivebestmdl_block(...)
}
\arguments{
\item{...}{Additional arguments passed to new_transform_block}
}
\description{
Displays the five best ARIMA models considered during seasonal adjustment.
Shows model specifications and their AICc values.
}
