% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_seasonal_blocks}
\alias{register_seasonal_blocks}
\title{Register seasonal adjustment blocks}
\usage{
register_seasonal_blocks()
}
\description{
Register all seasonal adjustment blocks with the blockr registry
}
